@@ -11,18 +11,21 @@
     "glob": "^7.1.7",
     "html-react-parser": "^1.4.10",
     "jquery": "^3.5.1",
+    "lightgallery": "^2.7.1",
     "list-files": "^1.1.4",
     "list-react-files": "^0.2.0",
     "mongodb": "^3.6.3",
     "mongoose": "^5.11.15",
+    "next-gallery": "^1.1.0",
     "nodemailer": "^6.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.2",
+    "react-dom": "^18.0.2",
     "react-lazyload": "^3.2.0",
-    "react-optimized-image": "^0.4.1",
+    "react-optimized-image": "^0.2.0",
+    "react-photo-gallery-next": "^8.0.1",
     "react-responsive-masonry": "^2.1.4",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.2",
+    "react-scripts": "^2.1.3",
     "react-spring": "^9.4.4",
     "rellax": "^1.12.1",
     "web-vitals": "^1.1.0"
